Time: 74 ms (7.69%), Space: 57.4 MB (27.05%) - LeetHub

+++ 2699-count-the-number-of-fair-pairs/2699-count-the-number-of-fair-pairs.java
@@ -7,7 +7,7 @@ class Solution {
             long low = lower - nums[i];
             long up = upper - nums[i];
 
-            cnt += (long)(search(i + 1, n - 1, up + 1, nums) - search(i + 1, n - 1, low, nums));
+            cnt += (search(i + 1, n - 1, up + 1, nums) - search(i + 1, n - 1, low, nums));
         }
         
         return cnt;
